feat(agent-tars-cli): implement based on tarko cli
