--- multimodal/agent-tars-cli/src/config/index.ts
@@ -1,2 +1,0 @@
-export * from './loader';
-export * from './builder';
