--- multimodal/agent-tars-cli/src/core/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './interactive-ui';
-export * from './headless-server';
-export * from './request';
-export * from './run';
-export * from './server-run';
-export * from './state';
-export * from './types';
