--- multimodal/agent-tars-cli/src/utils/index.ts
@@ -1,8 +1,0 @@
-/*
- * Copyright (c) 2025 Bytedance, Inc. and its affiliates.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export * from './logo';
-export * from './misc';
-export * from './console-interceptor';
